requirements.txt
fastapi[all]
uvicorn[standard]
pymongo[srv]
python-dotenv
